Backend/backend/requirements.txt
fastapi
uvicorn
python-dotenv
matplotlib
openpyxl
pyarrow
langchain_core
langgraph
langchain
langchain-openai
databricks-langchain
langchain_groq
graphviz
databricks-sql-connector
databricks-sdk
langchain-text-splitters
sentence-transformers
faiss-cpu
msal
pyjwt
python-jose
requests
pandas
python-multipart
